updated price to FLOAT instead of Int

+++ server/schemas/typeDefs.js
@@ -27,7 +27,7 @@ const typeDefs = gql`
         username: String
         image: String
         file: String
-        price: Int
+        price: Float
         difficulty: String
         category: Category
         reviews: [Review]
@@ -39,7 +39,7 @@ const typeDefs = gql`
         name: String
         username: String
         description: String
-        price: Int
+        price: Float
         classTime: String
         difficulty: String
         items: String
@@ -101,14 +101,14 @@ const typeDefs = gql`
         addOrder(classes:[ID]!, blueprints: [ID]!): Order
         login(email: String!, password: String!): Auth
         updateUser(firstName: String, lastName: String, email: String, password: String): User
-        addBlueprint(name: String!, description: String!, image: String, file: String, price: Int!, difficulty: String!, category: String!): Blueprint
-        addClass(name: String!, description: String!, price: Int!, classTime: String!, difficulty: String!, items: String!, category: String!): Class
+        addBlueprint(name: String!, description: String!, image: String, file: String, price: Float!, difficulty: String!, category: String!): Blueprint
+        addClass(name: String!, description: String!, price: Float!, classTime: String!, difficulty: String!, items: String!, category: String!): Class
         addPost(postText: String!, title: String!, category: String!): Post
         addCommentPost(postId: ID!, commentBody: String!): Post
         addClassReview(classId: ID!, reviewBody: String!): Class
         addBlueprintReview(blueprintId: ID!, reviewBody: String!): Blueprint
-        updateClass(classId: ID!, name: String!, description: String!, price: Int, items: String!): Class
-        updateBlueprint(blueprintId: ID!, name: String!, description: String!, price: Int): Blueprint
+        updateClass(classId: ID!, name: String!, description: String!, price: Float, items: String!): Class
+        updateBlueprint(blueprintId: ID!, name: String!, description: String!, price: Float): Blueprint
         updatePost(postId: ID!, title: String!, postBody: String!, category: String!): Post
         deleteBlueprint(blueprintId: ID!): Blueprint
         deleteClass(classId: ID!): Class
